abstract out createCheckboxControl
